Moving download into model.
the reason is that it can be used also for download data without a view.
